# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Sensible",
        "title": "Sensible Query Spec",
        "contact": {
            "name": "sensiblequery",
            "url": "https://github.com/sensible-contract/sensiblequery",
            "email": "jiedohh@gmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "2.0"
    },
    "paths": {
...
        "/address/{address}/contract-history": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "History"
                ],
                "summary": "通过地址address获取合约相关tx历史列表，返回详细输入/输出",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 666666,
                        "description": "Start Block Height",
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "End Block Height, (0 to get mempool data)",
                        "name": "end",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "起始游标",
                        "name": "cursor",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 16,
                        "description": "返回记录数量",
                        "name": "size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "17SkEw2md5avVNyYgj6RiXuQKNwkXaxFyQ",
                        "description": "Address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"data\": [{}], \"msg\": \"ok\"}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.TxOutHistoryResp"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/address/{address}/contract-history/tx": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "History"
                ],
                "summary": "通过地址address获取合约相关tx历史列表，返回tx概要",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 666666,
                        "description": "Start Block Height",
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "End Block Height, (0 to get mempool data)",
                        "name": "end",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "起始游标",
                        "name": "cursor",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 16,
                        "description": "返回记录数量",
                        "name": "size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "17SkEw2md5avVNyYgj6RiXuQKNwkXaxFyQ",
                        "description": "Address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"data\": [{}], \"msg\": \"ok\"}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.TxInfoResp"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/address/{address}/history": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "History"
                ],
                "summary": "通过地址address获取相关tx历史列表，返回详细输入/输出",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 666666,
                        "description": "Start Block Height",
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "End Block Height, (0 to get mempool data)",
                        "name": "end",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "起始游标",
                        "name": "cursor",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 16,
                        "description": "返回记录数量",
                        "name": "size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "17SkEw2md5avVNyYgj6RiXuQKNwkXaxFyQ",
                        "description": "Address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"data\": [{}], \"msg\": \"ok\"}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.TxOutHistoryResp"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/address/{address}/history/tx": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "History"
                ],
                "summary": "通过地址address获取相关tx历史列表，返回tx概要",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 666666,
                        "description": "Start Block Height",
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "End Block Height, (0 to get mempool data)",
                        "name": "end",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "起始游标",
                        "name": "cursor",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 16,
                        "description": "返回记录数量",
                        "name": "size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "17SkEw2md5avVNyYgj6RiXuQKNwkXaxFyQ",
                        "description": "Address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"data\": [{}], \"msg\": \"ok\"}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.TxInfoResp"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "model.TxInfoResp": {
            "type": "object",
            "properties": {
                "blkid": {
                    "description": "Tx所在区块的Id",
                    "type": "string"
                },
                "confirmations": {
                    "description": "Tx的确认数",
                    "type": "integer"
                },
                "height": {
                    "description": "Tx所在区块的高度",
                    "type": "integer"
                },
                "idx": {
                    "description": "Tx在区块中的顺序号",
                    "type": "integer"
                },
                "inSatoshi": {
                    "description": "Tx输入的satoshi总和",
                    "type": "integer"
                },
                "locktime": {
                    "description": "Tx Locktime",
                    "type": "integer"
                },
                "nIn": {
                    "description": "Tx包括的输入条数",
                    "type": "integer"
                },
                "nOut": {
                    "description": "Tx包括的输出条数",
                    "type": "integer"
                },
                "outSatoshi": {
                    "description": "Tx输出的satoshi总和",
                    "type": "integer"
                },
                "size": {
                    "description": "Tx字节数",
                    "type": "integer"
                },
                "timestamp": {
                    "description": "Tx所在区块的时间戳",
                    "type": "integer"
                },
                "txid": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}